// src/lib/supabase/types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "12.2.1 (d3f7cba)";
  };
  public: {
    Tables: {
// ...
      grades: {
        Row: {
          created_at: string | null;
          created_by: string | null;
          cycle_id: string;
          description: string;
          id: number;
          name: string;
          updated_at: string | null;
          updated_by: string | null;
        };
        Insert: {
          created_at?: string | null;
          created_by?: string | null;
          cycle_id: string;
          description: string;
          id?: number;
          name: string;
          updated_at?: string | null;
          updated_by?: string | null;
        };
        Update: {
          created_at?: string | null;
          created_by?: string | null;
          cycle_id?: string;
          description?: string;
          id?: number;
          name?: string;
          updated_at?: string | null;
          updated_by?: string | null;
        };
        Relationships: [
          {
            foreignKeyName: "grades_cycle_id_foreign";
            columns: ["cycle_id"];
            isOneToOne: false;
            referencedRelation: "cycles";
            referencedColumns: ["id"];
          },
        ];
      };
// ...
      schools: {
        Row: {
          address: string | null;
          city: string;
          code: string;
          created_at: string | null;
          created_by: string | null;
          cycle_id: string;
          email: string;
          id: string;
          image_url: string | null;
          is_technical_education: boolean | null;
          name: string;
          phone: string;
          state_id: number | null;
          status: Database["public"]["Enums"]["school_status_enum"];
          updated_at: string | null;
          updated_by: string | null;
        };
        Insert: {
          address?: string | null;
          city: string;
          code: string;
          created_at?: string | null;
          created_by?: string | null;
          cycle_id: string;
          email: string;
          id?: string;
          image_url?: string | null;
          is_technical_education?: boolean | null;
          name: string;
          phone: string;
          state_id?: number | null;
          status?: Database["public"]["Enums"]["school_status_enum"];
          updated_at?: string | null;
          updated_by?: string | null;
        };
        Update: {
          address?: string | null;
          city?: string;
          code?: string;
          created_at?: string | null;
          created_by?: string | null;
          cycle_id?: string;
          email?: string;
          id?: string;
          image_url?: string | null;
          is_technical_education?: boolean | null;
          name?: string;
          phone?: string;
          state_id?: number | null;
          status?: Database["public"]["Enums"]["school_status_enum"];
          updated_at?: string | null;
          updated_by?: string | null;
        };
        Relationships: [
          {
            foreignKeyName: "schools_cycle_id_foreign";
            columns: ["cycle_id"];
            isOneToOne: false;
            referencedRelation: "cycles";
            referencedColumns: ["id"];
          },
          {
            foreignKeyName: "schools_state_id_foreign";
            columns: ["state_id"];
            isOneToOne: false;
            referencedRelation: "states";
            referencedColumns: ["id"];
          },
        ];
      };
// ...
      student_school_class: {
        Row: {
          class_id: string | null;
          created_at: string | null;
          enrollment_status: string;
          grade_id: number;
          id: string;
          is_active: boolean;
          is_government_affected: boolean;
          is_orphan: boolean;
          is_redoublement: boolean;
          is_subscribed_to_canteen: boolean;
          is_subscribed_to_transportation: boolean;
          school_id: string;
          school_year_id: number;
          student_id: string;
          updated_at: string | null;
          updated_by: string | null;
        };
        Insert: {
          class_id?: string | null;
          created_at?: string | null;
          enrollment_status?: string;
          grade_id: number;
          id?: string;
          is_active?: boolean;
          is_government_affected?: boolean;
          is_orphan?: boolean;
          is_redoublement?: boolean;
          is_subscribed_to_canteen?: boolean;
          is_subscribed_to_transportation?: boolean;
          school_id: string;
          school_year_id: number;
          student_id: string;
          updated_at?: string | null;
          updated_by?: string | null;
        };
        Update: {
          class_id?: string | null;
          created_at?: string | null;
          enrollment_status?: string;
          grade_id?: number;
          id?: string;
          is_active?: boolean;
          is_government_affected?: boolean;
          is_orphan?: boolean;
          is_redoublement?: boolean;
          is_subscribed_to_canteen?: boolean;
          is_subscribed_to_transportation?: boolean;
          school_id?: string;
          school_year_id?: number;
          student_id?: string;
          updated_at?: string | null;
          updated_by?: string | null;
        };
        Relationships: [
          {
            foreignKeyName: "fk_class";
            columns: ["class_id"];
            isOneToOne: false;
            referencedRelation: "class_year_average_view";
            referencedColumns: ["class_id"];
          },
          {
            foreignKeyName: "fk_class";
            columns: ["class_id"];
            isOneToOne: false;
            referencedRelation: "classes";
            referencedColumns: ["id"];
          },
          {
            foreignKeyName: "fk_class";
            columns: ["class_id"];
            isOneToOne: false;
            referencedRelation: "student_conduct_summary_view";
            referencedColumns: ["class_id"];
          },
          {
            foreignKeyName: "fk_grade";
            columns: ["grade_id"];
            isOneToOne: false;
            referencedRelation: "grades";
            referencedColumns: ["id"];
          },
          {
            foreignKeyName: "fk_school";
            columns: ["school_id"];
            isOneToOne: false;
            referencedRelation: "schools";
            referencedColumns: ["id"];
          },
          {
            foreignKeyName: "fk_school_year";
            columns: ["school_year_id"];
            isOneToOne: false;
            referencedRelation: "average_grades_view_with_rank";
            referencedColumns: ["school_year_id"];
          },
          {
            foreignKeyName: "fk_school_year";
            columns: ["school_year_id"];
            isOneToOne: false;
            referencedRelation: "school_years";
            referencedColumns: ["id"];
          },
          {
            foreignKeyName: "fk_school_year";
            columns: ["school_year_id"];
            isOneToOne: false;
            referencedRelation: "student_semester_average_view";
            referencedColumns: ["school_year_id"];
          },
          {
            foreignKeyName: "fk_student";
            columns: ["student_id"];
            isOneToOne: false;
            referencedRelation: "payment_details_view";
            referencedColumns: ["student_id"];
          },
          {
            foreignKeyName: "fk_student";
            columns: ["student_id"];
            isOneToOne: false;
            referencedRelation: "payment_view";
            referencedColumns: ["student_id"];
          },
          {
            foreignKeyName: "fk_student";
            columns: ["student_id"];
            isOneToOne: false;
            referencedRelation: "student_conduct_summary_view";
            referencedColumns: ["student_id"];
          },
          {
            foreignKeyName: "fk_student";
            columns: ["student_id"];
            isOneToOne: false;
            referencedRelation: "student_payment_status_view";
            referencedColumns: ["student_id"];
          },
          {
            foreignKeyName: "fk_student";
            columns: ["student_id"];
            isOneToOne: false;
            referencedRelation: "students";
            referencedColumns: ["id"];
          },
        ];
      };
// ...
      user_roles: {
        Row: {
          grade_id: number | null;
          role_id: number;
          school_id: string | null;
          user_id: string;
        };
        Insert: {
          grade_id?: number | null;
          role_id: number;
          school_id?: string | null;
          user_id: string;
        };
        Update: {
          grade_id?: number | null;
          role_id?: number;
          school_id?: string | null;
          user_id?: string;
        };
        Relationships: [
          {
            foreignKeyName: "user_roles_grade_id_fkey";
            columns: ["grade_id"];
            isOneToOne: false;
            referencedRelation: "grades";
            referencedColumns: ["id"];
          },
          {
            foreignKeyName: "user_roles_role_id_fkey";
            columns: ["role_id"];
            isOneToOne: false;
            referencedRelation: "roles";
            referencedColumns: ["id"];
          },
          {
            foreignKeyName: "user_roles_school_id_fkey";
            columns: ["school_id"];
            isOneToOne: false;
            referencedRelation: "schools";
            referencedColumns: ["id"];
          },
          {
            foreignKeyName: "user_roles_user_id_fkey";
            columns: ["user_id"];
            isOneToOne: false;
            referencedRelation: "users";
            referencedColumns: ["id"];
          },
        ];
      };
      users: {
        Row: {
          avatar_url: string | null;
          created_at: string | null;
          created_by: string | null;
          email: string;
          first_name: string | null;
          id: string;
          last_name: string | null;
          phone: string | null;
          push_token: string | null;
          state_id: number | null;
          updated_at: string | null;
          updated_by: string | null;
        };
        Insert: {
          avatar_url?: string | null;
          created_at?: string | null;
          created_by?: string | null;
          email: string;
          first_name?: string | null;
          id?: string;
          last_name?: string | null;
          phone?: string | null;
          push_token?: string | null;
          state_id?: number | null;
          updated_at?: string | null;
          updated_by?: string | null;
        };
        Update: {
          avatar_url?: string | null;
          created_at?: string | null;
          created_by?: string | null;
          email?: string;
          first_name?: string | null;
          id?: string;
          last_name?: string | null;
          phone?: string | null;
          push_token?: string | null;
          state_id?: number | null;
          updated_at?: string | null;
          updated_by?: string | null;
        };
        Relationships: [
          {
            foreignKeyName: "users_state_id_foreign";
            columns: ["state_id"];
            isOneToOne: false;
            referencedRelation: "states";
            referencedColumns: ["id"];
          },
        ];
      };
    };
// ...
  };
};